# account_manager/openapi.yaml
openapi: 3.0.3
info:
  title: Account Manager Backend
  version: 1.0.0
  contact: {}
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /api/user/payment-account:
    get:
      parameters:
        - in: query
          name: offset
          schema:
            type: integer
          description: The number of items to skip before starting to collect the result set
        - in: query
          name: limit
          schema:
            type: integer
          description: The numbers of items to return
      tags:
        - Payment Accounts
      summary: List payment accounts in user
      description: List payment accounts in user
      operationId: listPaymentAccountsInUser
      responses:
        '200':
          description: List payment accounts in user
          headers:
            Connection:
              schema:
                type: string
                example: keep-alive
            Date:
              schema:
                type: string
                example: Sat, 11 May 2024 06:17:14 GMT
            Keep-Alive:
              schema:
                type: string
                example: timeout=72
            access-control-allow-credentials:
              schema:
                type: string
                example: 'true'
            access-control-allow-origin:
              schema:
                type: string
                example: http://localhost:8080
            content-length:
              schema:
                type: string
                example: '105'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Success
                  result:
                    type: array
                    items:
                      type: object
                      properties:
                        amount:
                          type: number
                          example: 7457.85
                        currency:
                          type: string
                          example: JPY
                        id:
                          type: string
                          example: '1715407506'
                        type:
                          type: string
                          example: Debit JPY
                    example:
                      - amount: 7457.85
                        currency: JPY
                        id: '1715407506'
                        type: Debit JPY
              examples:
                List payment accounts in user:
                  value:
                    message: Success
                    result:
                      - amount: 7457.85
                        currency: JPY
                        id: '1715407506'
                        type: Debit JPY
      security:
        - bearerAuth: []
    post:
      tags:
        - Payment Accounts
      summary: Create payment account
      description: Create payment account
      operationId: createPaymentAccount
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                currency:
                  type: string
                  example: JPY
                type:
                  type: string
                  example: Debit JPY
            examples:
              Create payment account:
                value:
                  currency: JPY
                  type: Debit JPY
      responses:
        '200':
          description: Create payment account
          headers:
            Connection:
              schema:
                type: string
                example: keep-alive
            Date:
              schema:
                type: string
                example: Sat, 11 May 2024 05:48:40 GMT
            Keep-Alive:
              schema:
                type: string
                example: timeout=72
            access-control-allow-credentials:
              schema:
                type: string
                example: 'true'
            access-control-allow-origin:
              schema:
                type: string
                example: http://localhost:8080
            content-length:
              schema:
                type: string
                example: '66'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Created payment account
                  result:
                    type: object
                    properties:
                      id:
                        type: string
                        example: '1715406521'
              examples:
                Create payment account:
                  value:
                    message: Created payment account
                    result:
                      id: '1715406521'
      security:
        - bearerAuth: []
...
  /api/user/payment-account/{id}:
    get:
      tags:
        - Payment Accounts
      summary: Get payment account by id
      description: Get payment account by id
      operationId: getPaymentAccountById
      responses:
        '200':
          description: Get payment account by id
          headers:
            Connection:
              schema:
                type: string
                example: keep-alive
            Date:
              schema:
                type: string
                example: Sat, 11 May 2024 06:17:51 GMT
            Keep-Alive:
              schema:
                type: string
                example: timeout=72
            access-control-allow-credentials:
              schema:
                type: string
                example: 'true'
            access-control-allow-origin:
              schema:
                type: string
                example: http://localhost:8080
            content-length:
              schema:
                type: string
                example: '103'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Success
                  result:
                    type: object
                    properties:
                      amount:
                        type: number
                        example: 7457.85
                      currency:
                        type: string
                        example: JPY
                      id:
                        type: string
                        example: '1715407506'
                      type:
                        type: string
                        example: Debit JPY
              examples:
                Get payment account by id:
                  value:
                    message: Success
                    result:
                      amount: 7457.85
                      currency: JPY
                      id: '1715407506'
                      type: Debit JPY
      security:
        - accessTokenCookie: []
        - accessTokenBearer: []
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          example: '1715407506'
...
components:
  securitySchemes:
    accessTokenBearer:
      description: Returned as the "st-access-token" header from sign-in and refresh endpoints and present on requests that update the access token payload
      type: http
      scheme: bearer
    refreshTokenBearer:
      description: Returned as the "st-refresh-token" header from sign-in and refresh endpoints
      scheme: bearer
      type: http
    accessTokenCookie:
      type: apiKey
      in: cookie
      name: sAccessToken  
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: sRefreshToken  
...